se agrego un poco de modularizacion a RoomLobby
